requirements.txt
Django==5.1.3
django-filter==24.1
asgiref==3.8.1
cffi==1.16.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gevent==23.7.0
greenlet==3.0.3
gunicorn==23.0.0
openpyxl==3.1.2
packaging==24.0
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
zope.event==5.0
zope.interface==6.2
icecream==2.1.3
pandas==2.2.3
numpy==2.1.3
pytest==8.3.3
